pyproject.toml
# Constitutional Hash: 608508a9bd224290
[build-system]
requires = ["setuptools>=80.9.0"]
build-backend = "setuptools.build_meta"

[project]
name = "acgs"
version = "3.0.1"
description = "Partial-open-source ACGS package line for governance runtime, shared components, and platform tooling"
requires-python = ">=3.11"
license = "AGPL-3.0-or-later"
dependencies = [
    "fastapi>=0.127.0",
    "httpx>=0.27.0",
    "pydantic>=2.12.1",
    "pydantic-settings>=2.13.1",
    "redis>=5.0.0",
    "structlog>=25.5.0",
    "prometheus-client>=0.20.0",
    "PyJWT>=2.8.0",
    "orjson>=3.10.0",
    "cachetools>=5.0.0",
    "pybreaker>=1.4.1",
    "slowapi>=0.1.9",
    "msgpack>=1.0.0",
    "email-validator>=2.3.0",
    "python-multipart>=0.0.22",
    "uvloop>=0.21.0",
    "httptools>=0.6.0",
    "PyYAML>=6.0.0",
    "psutil>=5.9.0",
    "jsonschema>=4.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]
test = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-xdist>=3.3.1",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.4.0",
    "pytest-mock>=3.11.0",
    "pytest-benchmark>=4.0.0",
    "fakeredis>=2.18.0",
    "hypothesis>=6.151.9",
    "respx>=0.20.0",
    "aiosqlite>=0.20.0",
    "coverage[toml]>=7.3.0",
]
ml = [
    "numpy>=1.24.0",
    "scikit-learn>=1.4.0",
    "pandas>=2.1.0",
]
pqc = [
    "liboqs-python>=0.14.1",
    "cryptography>=44.0.2",
]
postgres = [
    "asyncpg>=0.31.0",
    "sqlalchemy>=2.0.38",
]

[tool.uv]
managed = true
package = true

[tool.uv.workspace]
members = ["packages/enhanced_agent_bus", "packages/acgs-lite", "packages/acgs-deliberation"]

[tool.uv.sources]
enhanced-agent-bus = { workspace = true }
acgs-lite = { workspace = true }
acgs-deliberation = { workspace = true }

[tool.ruff]
line-length = 100
target-version = "py311"
exclude = [".claude", ".codex", ".codex-home", ".codex-readiness-integration-test", "demo", "examples", "node_modules", "__pycache__", "tools"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "BLE", "S", "RUF", "C90"]
ignore = [
    "E501",
    "B008",
    "S101",
    "E402",
    "C901",
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S105", "S106", "S108", "BLE001", "S311", "F811", "B017", "S110", "B905"]
"**/tests/**/*.py" = ["S105", "S106", "S108", "BLE001", "S311", "F811", "B017", "S110", "B905"]
"packages/enhanced_agent_bus/**/*.py" = ["BLE001", "S311"]
"packages/acgs-lite/**/*.py" = ["BLE001", "S311"]
"packages/acgs-core/**/*.py" = ["BLE001", "S102"]
"packages/clinicalguard/**/*.py" = ["BLE001", "S104", "S108", "S110"]
"packages/constitutional_swarm/**/*.py" = ["S105"]
"src/**/*.py" = ["BLE001", "S105", "S106", "S108", "RUF001", "RUF002", "RUF003"]
"src/acgs_lite/**/*.py" = ["BLE001", "S311", "F401", "RUF012"]
"autoresearch/**/*.py" = ["BLE001", "S311", "T201"]
"**/__init__.py" = ["F401"]
"**/examples/**/*.py" = ["E402", "S105", "S106", "BLE001"]
"packages/acgs-lite/examples/*.py" = ["E402", "S105", "S106", "BLE001"]

[tool.ruff.lint.mccabe]
max-complexity = 15

[tool.ruff.lint.isort]
known-first-party = ["src", "packages", "acgs_lite", "enhanced_agent_bus", "acgs_deliberation", "clinicalguard"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
exclude = ["build", "dist", ".codex-home", ".venv", "tests"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "--import-mode=importlib"
pythonpath = ["."]
testpaths = [
    "tests",
    "packages/enhanced_agent_bus",
    "packages/acgs-lite/tests",
    "packages/acgs-deliberation/tests",
    "src/core/shared",
    "src/core/services/api_gateway/tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
asyncio_mode = "auto"
markers = [
    "unit: marks tests as unit tests",
    "integration: marks tests as integration tests",
    "slow: marks tests as slow",
    "constitutional: marks tests for constitutional policy validation",
    "benchmark: marks performance benchmark tests",
    "governance: marks governance critical-path tests",
    "security: marks security validation tests",
    "maci: marks MACI enforcement tests",
    "chaos: marks chaos engineering tests",
    "pqc: marks post-quantum cryptography tests",
    "pqc_deprecation: marks PQC deprecation and migration tests",
    "compliance: marks compliance validation tests",
    "e2e: marks end-to-end tests",
    "postgres: marks tests requiring PostgreSQL-specific behavior",
]

[tool.coverage.run]
source = ["src", "packages/enhanced_agent_bus", "packages/acgs-lite/src", "packages/acgs-deliberation/src"]
omit = ["tests/*", "**/tests/*", "**/__init__.py", "src/acgs_lite/*"]

[tool.coverage.report]
fail_under = 70
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@(abc\\.)?abstractmethod",
]
